# tests/fixtures/t41_http_contract/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_create_template_api_v1_templates_post": {
        "properties": {
          "content": {
            "contentMediaType": "application/octet-stream",
            "title": "Content",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "expected_fonts": {
            "items": {
              "type": "string"
            },
            "title": "Expected Fonts",
            "type": "array"
          },
          "kind": {
            "$ref": "#/components/schemas/TemplateKind",
            "default": "docx"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "description",
          "expected_fonts",
          "content"
        ],
        "title": "Body_create_template_api_v1_templates_post",
        "type": "object"
      },
...
      "ErrorDetail": {
        "description": "Stable machine-readable functional error detail.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "Stable error envelope used for every expected HTTP failure.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "TemplateAdministrationContextResponse": {
        "description": "Authoritative template selection identifiers and upload limit.",
        "properties": {
          "preferred_template_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Template Id"
          },
          "system_fallback_template_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System Fallback Template Id"
          },
          "template_max_archive_bytes": {
            "exclusiveMinimum": 0.0,
            "title": "Template Max Archive Bytes",
            "type": "integer"
          }
        },
        "required": [
          "preferred_template_id",
          "system_fallback_template_id",
          "template_max_archive_bytes"
        ],
        "title": "TemplateAdministrationContextResponse",
        "type": "object"
      },
      "TemplateKind": {
        "description": "Immutable reference document family.",
        "enum": [
          "docx",
          "pptx"
        ],
        "title": "TemplateKind",
        "type": "string"
      },
...
      "TemplatePageResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TemplateResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "TemplatePageResponse",
        "type": "object"
      },
...
      "TemplateStatus": {
        "description": "Visibility lifecycle available before T15 adds version mutations.",
        "enum": [
          "active",
          "archived"
        ],
        "title": "TemplateStatus",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "description": "Opaque authenticated Markweave session cookie.",
        "in": "cookie",
        "name": "md_converter_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Markdown Converter API",
    "version": "0.7.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/template-context": {
      "get": {
        "operationId": "get_template_administration_context_api_v1_template_context_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateAdministrationContextResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication failed or is required",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The service is not ready",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Get Template Administration Context",
        "tags": [
          "templates"
        ]
      }
    },
    "/api/v1/template-preference": {
      "delete": {
        "operationId": "clear_preferred_template_api_v1_template_preference_delete",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication failed or is required",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The operation is forbidden",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The request is invalid",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The service is not ready",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Clear Preferred Template",
        "tags": [
          "templates"
        ]
      }
    },
    "/api/v1/templates": {
      "get": {
        "operationId": "list_templates_api_v1_templates_get",
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            }
          },
          {
            "in": "query",
            "name": "description",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Description"
            }
          },
          {
            "in": "query",
            "name": "owner_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TemplateStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TemplateKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplatePageResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication failed or is required",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The request is invalid",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The service is not ready",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "List Templates",
        "tags": [
          "templates"
        ]
      },
      "post": {
        "operationId": "create_template_api_v1_templates_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_template_api_v1_templates_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "ETag": {
                "description": "Current template identity or immutable content validator.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication failed or is required",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The operation is forbidden",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The request body is too large",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The request is invalid",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The service is not ready",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Create Template",
        "tags": [
          "templates"
        ]
      }
    },
...
  },
  "security": [
    {
      "SessionCookie": []
    }
  ]
}